MARS-1015 Restructure `client`
